add file logging to commands `full-install` and `update apply`
